move unrelated functions out of thegraph module
